+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
+## v 1.1.0 (2024-01-09)
+
+-   fix: Fix crashing bug.
+-   fix: Get BrowserWindow and port `_createWindow()` method.
+-   feature: Port htmlContent.
+
+---
+
 ## v 0.0.1 (2024-01-08)
 
 -   feature: [Add new `CHANGELOG.md`](/CHANGELOG.md)
--   feature:
-    -   feature
-        -   electron-progressbar porting.
+-   feature: Port electron-progressbar.